requirements.txt
PySide6==6.7.2
supabase>=2.5
python-dotenv>=1.0
pandas>=2.2
openpyxl>=3.1
requests>=2.31
PyJWT>=2.8
